Fixed code-smells issues, added test coverage jacoco plugin

+++ src/main/java/bit/wcservice/web/service/CurrencyWebService.java
@@ -2,5 +2,5 @@ package bit.wcservice.web.service;
 
 public interface CurrencyWebService {
     String loadCurrentUSDValue();
-    String loadLastDaysUSDHistory(int days);
+    String loadLastDaysUSDHistory(long days);
 }
