+++ src/main/java/bit/wcservice/web/service/WeatherWebService.java
@@ -2,5 +2,5 @@ package bit.wcservice.web.service;
 
 public interface WeatherWebService {
     String loadCurrentWeatherIn(String location);
-    String loadLastDaysWeatherHistoryInLocation(int days, String location);
+    String loadLastDaysWeatherHistoryInLocation(long days, String location);
 }
